moved NetworkConfigOperations to generic.chain package
