+++ src/main/scala/tech/cryptonomic/conseil/routes/PlatformDiscovery.scala
@@ -4,7 +4,8 @@ import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.Route
 import com.typesafe.config.Config
 import com.typesafe.scalalogging.LazyLogging
-import tech.cryptonomic.conseil.tezos.{ApiNetworkOperations, NetworkConfigOperations, PlatformDiscoveryOperations}
+import tech.cryptonomic.conseil.generic.chain.NetworkConfigOperations
+import tech.cryptonomic.conseil.tezos.{ApiNetworkOperations, PlatformDiscoveryOperations}
 import tech.cryptonomic.conseil.util.JsonUtil._
 import tech.cryptonomic.conseil.util.RouteHandling
